Fix py312 random int range
Python 3.12 tests were failing due to the fact that we tried to create a
random int using a range with floats.

+++ mistral/scheduler/default_scheduler.py
@@ -85,7 +85,8 @@ class DefaultScheduler(base.Scheduler):
 
             eventlet.sleep(
                 self._fixed_delay +
-                random.Random().randint(0, self._random_delay * 1000) * 0.001
+                random.Random().randint(
+                    0, int(self._random_delay * 1000)) * 0.001
             )
 
             try:
